[Bug] delete the default app bug fixed #2252 (#2280)

+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/service/impl/ApplicationServiceImpl.java
@@ -34,6 +34,7 @@ import org.apache.streampark.common.util.HadoopUtils;
 import org.apache.streampark.common.util.ThreadUtils;
 import org.apache.streampark.common.util.Utils;
 import org.apache.streampark.common.util.YarnUtils;
+import org.apache.streampark.console.base.domain.Constant;
 import org.apache.streampark.console.base.domain.RestRequest;
 import org.apache.streampark.console.base.exception.ApiAlertException;
 import org.apache.streampark.console.base.exception.ApiDetailException;
@@ -369,41 +370,51 @@ public class ApplicationServiceImpl extends ServiceImpl<ApplicationMapper, Appli
 
     Application application = getById(paramApp.getId());
 
-    try {
-      // 1) remove flink sql
-      flinkSqlService.removeApp(application.getId());
-
-      // 2) remove log
-      applicationLogService.removeApp(application.getId());
-
-      // 3) remove config
-      configService.removeApp(application.getId());
-
-      // 4) remove effective
-      effectiveService.removeApp(application.getId());
-
-      // remove related hdfs
-      // 5) remove backup
-      backUpService.removeApp(application);
-
-      // 6) remove savepoint
-      savePointService.removeApp(application);
-
-      // 7) remove BuildPipeline
-      appBuildPipeService.removeApp(application.getId());
-
-      // 8) remove app
-      removeApp(application);
-
-      if (isKubernetesApp(paramApp)) {
-        k8SFlinkTrackMonitor.unWatching(toTrackId(application));
+    if (application.getExecutionMode() == null) {
+      if (application.getId().equals(Constant.FLINK_SAMPLE_APP_ID)) {
+        flinkSqlService.removeApp(application.getId());
+        removeById(application.getId());
+        return true;
       } else {
-        FlinkRESTAPIWatcher.unWatching(paramApp.getId());
+        throw new ApiAlertException("job is invalid, executionMode is null");
       }
-      return true;
-    } catch (Exception e) {
-      log.error(e.getMessage(), e);
-      throw e;
+    } else {
+      try {
+        // 1) remove flink sql
+        flinkSqlService.removeApp(application.getId());
+
+        // 2) remove log
+        applicationLogService.removeApp(application.getId());
+
+        // 3) remove config
+        configService.removeApp(application.getId());
+
+        // 4) remove effective
+        effectiveService.removeApp(application.getId());
+
+        // remove related hdfs
+        // 5) remove backup
+        backUpService.removeApp(application);
+
+        // 6) remove savepoint
+        savePointService.removeApp(application);
+
+        // 7) remove BuildPipeline
+        appBuildPipeService.removeApp(application.getId());
+
+        // 8) remove app
+        removeApp(application);
+
+        if (isKubernetesApp(paramApp)) {
+          k8SFlinkTrackMonitor.unWatching(toTrackId(application));
+        } else {
+          FlinkRESTAPIWatcher.unWatching(paramApp.getId());
+        }
+        return true;
+      } catch (Exception e) {
+        log.error(e.getMessage(), e);
+        throw e;
+      }
     }
   }
